+++ SignNow.Net/Interfaces/IEventSubscriptionService.cs
@@ -87,5 +87,12 @@ namespace SignNow.Net.Interfaces
         /// <param name="cancellationToken">Propagates notification that operations should be canceled.</param>
         /// <returns>Events History page</returns>
         Task<EventHistoryListResponse> GetEventHistoryAsync(string eventId, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Allows to update an existing event subscription.
+        /// </summary>
+        /// <param name="updateEvent">Event details for update</param>
+        /// <param name="cancellationToken">Propagates notification that operations should be canceled.</param>
+        Task EditEventSubscriptionAsync(UpdateEventSubscription updateEvent, CancellationToken cancellationToken = default);
     }
 }